Primer commit dela clase de aprender a usar git
Though not required, it's a good idea to begin the commit message with a single short (less than 50 character) line summarizing the change,
followed by a blank line and then a more thorough description.
The text up to the first blank line in a commit message is treated as the commit title, and that title is used throughout Git.
For example, Git-format-patch(1) turns a commit into email, and it uses the title on the Subject line and the rest of the commit in the body.